+++ src/renderer/extensions/setting/index.less
@@ -1,5 +1,6 @@
 .lightproxy-setting {
     padding: 50px;
+    margin-top: 30px;
     line-height: 25px;
     font-weight: bold;
     font-size: 16px;
